[FIX] Команды и права.

+++ src/main/java/ru/zaxar163/indexer/command/manage/RemoveMsgCommand1.java
@@ -4,23 +4,22 @@ import java.util.Iterator;
 import java.util.stream.Stream;
 
 import org.javacord.api.entity.message.Message;
+import org.javacord.api.entity.permission.PermissionType;
 
 import ru.zaxar163.indexer.Indexer;
 import ru.zaxar163.indexer.RoleManager;
 import ru.zaxar163.indexer.command.Command;
 
-public class RemoveMsgCommand extends Command {
+public class RemoveMsgCommand1 extends Command {
 
-	private final Indexer indexer;
-
-	public RemoveMsgCommand(final Indexer indexer) {
+	public RemoveMsgCommand1(final Indexer indexer) {
 		super("removemsg", "!removemsg - удаляет сообщения в диапазоне");
-		this.indexer = indexer;
+		super.aliases.add("rmd");
 	}
 
 	@Override
 	public boolean canUse(final Message message) {
-		return RoleManager.hasRole(indexer.roler.middleDeveloper, message.getUserAuthor());
+		return super.canUse(message) && RoleManager.hasAnyPerm(message, PermissionType.ADMINISTRATOR, PermissionType.MANAGE_MESSAGES);
 	}
 
 	@Override
@@ -29,10 +28,11 @@ public class RemoveMsgCommand extends Command {
 		if (args.length < 1)
 			throw new IllegalArgumentException("Invalid args.");
 		final int cnt = Integer.parseInt(args[0]);
+		if (cnt < 1) throw new IllegalArgumentException("Invalid args.");
 		final Iterator<Message> it = msgs.iterator();
 		if (!it.hasNext())
 			return;
-		for (int i = 0; i < cnt && it.hasNext(); i++)
+		for (int i = 0; i <= cnt && it.hasNext(); i++)
 			it.next().delete();
 		msgs.close();
 	}
